drivers/pty: Close "slave" TTYs when the "master" is closed.
The old TTY driver did this, but it was not properly carried
over into this new driver.

Fixes the first part of #18488: the child bash process from the
first session is properly closed, but the TTY state isn't reset
and there can still be stale things left in its buffer, it appears.