network/protocols/unix: Fix UnixDatagramEndpoint::Close().
1. Set fShutdownRead & fShutdownWrite. This is necessary
because in Send(), there is a lock-unlock dance between the
source and target sockets which could potentially race with
some other thread closing the socket. The "shutdown" state
is checked before actually writing, and so we need to set it
here. Otherwise, the other thread could allocate a new receive
FIFO for us, which we do not want.

2. Don't return early when unbinding, but only if there's an
error (there shouldn't be.)

3. Merely ASSERT() in Free() that closing was already done
and nothing remains to be deallocated.

Should fix #18535.
unix_dgram_test still passes.