Icon-O-Matic: Improve saving-related error messages
Improves the errors that are given to the user when exporting to an
HVIF file failed. Examples of errors improved include errors related to
having too many shapes, too many paths, or too many points in a path.

Also changes the icon shown in the alert box from an info icon to an
error icon.

Fixes #13978

Change-Id: I12a08e9857a95bd4d41da029320fa6f8182c3aa5
Reviewed-on: https://review.haiku-os.org/c/haiku/+/6837
Reviewed-by: Adrien Destugues <pulkomandy@pulkomandy.tk>
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>