ShowImage: Add "Open with..." menu
Populate a new sub-menu with all applications that support the current
image's MIME type. Launch the chosen app with the image entry_ref as
B_REFS_RECEIVED message.

The sub-menu listing supporting apps is separated into apps supporting
the supertype ("image/" and full support (e.g. "image/png").

Add the new "Open with..." menu to the "File" menu and the context menu
of the image.

Most of the code was grabbed from the FileType prefs PreferredAppMenu,
therefore its copyright was copied and extended.

Change-Id: I0190cf4b78c59e429f4e67654665598340653c5d
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7072
Reviewed-by: Axel Dörfler <axeld@pinc-software.de>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: waddlesplash <waddlesplash@gmail.com>