People: Accept initial values
Besides the entry_ref to an existing Person file as argv, also
accept a BMessage containing values for attributes and set those
in the new window.

The specifier name in the BMessage needs to be the same as the
attribute name used by the People app, e.g. "META:email".

Needed to let the Mail app create a new Person file with the email
address already filled in.

Fixes #14766

Change-Id: I637200b08d5c1b7cf52d7cf3f56dab93ce1a6ec5
Reviewed-on: https://review.haiku-os.org/c/haiku/+/7085
Reviewed-by: Adrien Destugues <pulkomandy@pulkomandy.tk>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>